feat(consensus/types): move StampingVote to package core
